style: Code refactored back and tests added
