+++ courses/tasks_test.py
@@ -2,22 +2,20 @@
 
 import pytest
 
-from courses.tasks import generate_course_certificates, sync_courseruns_data
+from courses.tasks import sync_courseruns_data
+from courses.factories import CourseFactory, CourseRunFactory
 
-pytestmark = pytest.mark.django_db
-
-
-def test_generate_course_certificates_task(mocker):
-    """Test generate_course_certificates calls the right api functionality from courses"""
-    generate_course_run_certificates = mocker.patch(
-        "courses.api.generate_course_run_certificates"
-    )
-    generate_course_certificates.delay()
-    generate_course_run_certificates.assert_called_once()
+pytestmark = [pytest.mark.django_db]
 
 
 def test_sync_courseruns_data(mocker):
     """Test sync_courseruns_data calls the right api functionality from courses"""
-    sync_course_runs_data_task = mocker.patch("courses.api.sync_course_runs_data")
+    sync_course_runs = mocker.patch("courses.utils.sync_course_runs")
+
+    course_runs = CourseRunFactory.create_batch(size=3)
+    external_course_runs = CourseRunFactory.create_batch(
+        size=3, course__is_external=True
+    )
+
     sync_courseruns_data.delay()
-    sync_course_runs_data_task.assert_called_once()
+    sync_course_runs.assert_called_once_with(course_runs)
